+++ bind-events/1-finish/src/app/app.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-root',
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.scss']
+})
+export class AppComponent {
+  text: string = 'Default text'
+
+  changeText() {
+    this.text = 'Custom text'
+  }
+}