+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### :star: Added
 - exercise `generate-angular-project` - generate Angular project
 - exercise `generate-component` - generate component
+- exercise `bind-events` - bind events
 
 ### :hammer: Changed